refactor(test): split oversized test files to meet LOC limits
Split test_run_supervisor.py (760 -> 150 lines) into test_run_manager_issue.py.
Split test_codeagent_backend.py (409 -> 252 lines) into test_codeagent_async.py.
